fix: rename form-blcok-type-icon.tsx to form-block-type-icon.tsx
