+++ editor/scaffolds/sidebar/sidebar-mode-design-form.tsx
@@ -8,7 +8,7 @@ import type {
   FormBlockType,
   FormInputType,
 } from "@/grida-forms-hosted/types";
-import { BlockTypeIcon } from "@/components/form-blcok-type-icon";
+import { BlockTypeIcon } from "@/components/form-block-type-icon";
 import { FormFieldTypeIcon } from "@/components/form-field-type-icon";
 import { DotsHorizontalIcon } from "@radix-ui/react-icons";
 import {
