+++ editor/scaffolds/sidebar/sidebar-mode-insert.tsx
@@ -14,7 +14,7 @@ import {
 import FormField from "@/components/formfield/form-field";
 import { Button } from "@/components/ui/button";
 import { blocklabels } from "@/k/supported_block_types";
-import { BlockTypeIcon } from "@/components/form-blcok-type-icon";
+import { BlockTypeIcon } from "@/components/form-block-type-icon";
 import type { FormBlockType, FormInputType } from "@/grida-forms-hosted/types";
 import { SearchInput } from "@/components/extension/search-input";
 import { DummyFormAgentStateProvider } from "@/grida-forms/formstate";
